use consistent path var names in DehydrateTests
We try to use consistent variable names for all
variables in DehydrateTests which contain full
virtual paths as constructed by GetVirtualPathTo().

While this is not immediately helpful, it will be
crucial when we introduce the distinction between
virtual paths (full paths to the mounted test repo)
and backing root paths (full paths to the repo's
backing store which may be accessed while the repo
is unmounted on Linux).