Ensure git-credential cmds are run with correct wrkdir
Ensure that all git-credential commands are run with the correct working
directory. Currently only 'fill' is invoked against the .git directory
and the 'reject' and 'approve' commands are invoked outside the
repository.

The result of this mismatch means that if the user has a local
configuration entry set, for say credential.usehttppath, then the
behaviour for locating credentials will be different between read and
modification operations.