Create shared cache folders with the same ACLs as repo folders
The changes in this commit address the ACL problems that can
occur on Windows Server when a shared cache created by an
elevated VFS4G process is used by a non-elevated process.

By adding Authenticated Users to the shared cache ACLs we ensure
that the permissions of the shared cache will be consistent with
those used in the repo itself.