Update sparse parent folder mtimes to trigger Explorer refresh
If a File Explorer window is open to a directory in VFS4G, changes
to the files being projected are not enough to trigger a refresh.
This is especially noticeable when using the 'gvfs sparse' verb.

To force Explorer to refresh when 'gvfs sparse' triggers a projecion
changes, VFS4G will now update the last write time for every folder
placeholder that had children added or removed from the sparse set.