Change how RefreshSparseFolders discovers folders to refresh
The old approach (refreshing just the parent of a new path) could
miss some folders that need a refresh.

Example:

1. gvfs sparse --set A;B\C\D --prune
2. gvfs sparse --set A;B\D\F --prune

In the above example, B should be refreshed, but it would not be.

Rather than just refreshing the parent VFS4G will now refresh all
ancestors of entries that were added or removed from the sparse set.

This might result in over-refreshing, but we should no longer miss
any folders.