Refresh the parents of pruned directories
In addition to refreshing the parents of directories that were added
or removed from the sparse set, parents of pruned directories
must also have their write times updated or they might not properly
refresh in explorer.

Example (where refreshed pruned parent is required):

  1. Sparse set is A\B
  2. Create a new directory A\C
  3. gvfs --sparse prune (or gvfs sparse --set A\B --prune)

If A is not refreshed it will appear empty in Explorer as the last
refresh will have occured when A\C was deleted (while VFS4G was
unmounted) making A appear empty.