HandleDehydrateFolders: don't parse index a second time
HandleDehydrateFolders was triggering two parsese of the index:

 - Once in response to 'git reset' to validate the modified paths
   are correct
 - Once with an explicit call to ForceIndexProjectionUpdate

Making matters worse, the explicit call to ForceIndexProjectionUpdate
was not happening inside of an external GVFSLock, and so it could
collide with the parse triggered by reset mixed (which runs on the
background thread).

To prevent this race (and double parsing), HandleDehydrateFolders
now sets a flag 'resetForDehydrateInProgress' to true when it calls
'git reset', and when that flag is true the ForceIndexProjectionUpdate
that is called by HandlePostIndexChangedRequest will request that the
projection be updated.