GVFSVerb: use flags for core.gvfs
The core.gvfs config option can take a boolen "true" to turn on
all options, but also an integer to show a bitmask of important
flags. For the first time, we have a new flag that we may not
want to enable right away: GVFS_PREFETCH_DURING_FETCH.

This flag will make the `git fetch` builtin run a `git gvfs-helper
prefetch` command. Until that hardens a bit, we do not want VFS
for Git users to use that code path.

Update Git to include this option.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>