RepoRegistry: ignore non-existent repos
The 'gvfs mount' command will send a message over the named pipe to
deactivate a repo. This marks it with "IsActive = false" so we don't
try to mount it at service startup. However, a repo can be deleted
and then there is no way to send this message again.

Instead, ignore repositories where the enlistment root does not
exist. Don't remove it from the registry, since there may be a
subtle reason why it is not visible from the serivce (such as a
removed USB drive).

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>