ResetMixedTests: relax output checking
The sparse-checkout feature was updated with a more detailed warning
message. This only happens when a formerly-untracked file hits a file on
disk that used to be outside the sparse scope.

Since our control repo doesn't use sparse-checkout but VFS for Git uses
the virtualization hook to feed into the sparse-checkout code, this
means we need to relax the conditions on our test.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>