Config: disable status.submoduleSummary
The GVFS protocol does not support submodules, so disable the
status.submoduleSummary config setting. A user had this enabled
in their global config, but that caused "git status" to be much
slower. When enabled, this runs two "git submodule" subprocesses,
which are very slow.

I investigated disabling these subprocesses in Git when the repo
has no submodules, but it needs to run the processes to check for
untracked submodules. Thus, there is no efficient way to detect
that submodules exist or not in advance.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>