GVFS.SignFiles.csproj: include more DLLs for signing
It appears that during our build, we send some DLLs for signing, but the
versions we send have already been copied to other projects. Thus, we
ship versions of those DLLs that are unsigned.

Add extra versions of the DLLs that are currently being signed, except
include them in the GVFS.Windows project that is likely where they are
being copied as part of the build.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>