GitObjects: hash incoming loose objects
After we write the data to disk, check the file that it matches the
correct hash before moving it into place. If the hash is wrong, then
include details about where the data is stored for reference.

This can be marked as optional to make existing unit tests work, but the
logic is covered by a new set of unit tests.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>