GitObjects: use new SideChannelStream to hash during download
The loose object download needs to verify the objects being sent across
the wire before moving them to their "real" locations in the object
directory. The previous implementation wrote them to disk then read from
that file to hash the contents.

Instead, create a new "SideChannelStream" that injects between the
responseStream and the inflater, but sends all of the data that is read
out to the filestream. Here is a diagram of the stream sequence:

  responseStream -> sideChannelStream -> inflateStream -> hashingStream
                           \                                   |
			    \                                 \|/
			     tempFileStream                devNull

By copying all data from the hashingStream into devNull, we ensure that
the responseStream is drained and written to the tempFileStream while
simultaneously inflating and hashing the object contents. We then
amortize this computation during the window that we are downloading data
over the network, minimizing the potential performance impact.

Helped-by: Matthew Cheetham <mjcheetham@github.com>
Signed-off-by: Derrick Stolee <dstolee@microsoft.com>