GVFSVerb: strip commit id from version, if it exists
When trying to run functional tests for a locally-build version, I got
the following error:

Cannot clone @ C:\Repos\GVFSFunctionalTests\enlistment\29fd71cb642b4523bc0a: System.FormatException: Input string was not in a correct format.
   at System.Version.VersionResult.SetFailure(ParseFailureKind failure, String argument)
   at System.Version.TryParseComponent(String component, String componentName, VersionResult& result, Int32& parsedComponent)
   at System.Version.TryParseVersion(String version, VersionResult& result)
   at System.Version.Parse(String input)
   at System.Version..ctor(String version)
   at GVFS.CommandLine.GVFSVerb.TryValidateGVFSVersion(GVFSEnlistment enlistment, ITracer tracer, ServerGVFSConfig config, String& errorMessage, Boolean& errorIsFatal)
   at GVFS.CommandLine.GVFSVerb.ValidateClientVersions(ITracer tracer, GVFSEnlistment enlistment, ServerGVFSConfig gvfsConfig, Boolean showWarnings)
   at GVFS.CommandLine.CloneVerb.Execute()

The version string I was seeing was of the form "GVFS 0.2.XXX.Y+<id>"
so my guess is that the "+<id>" portion is causing the System.Version
logic to complain.