Update the LibGit2Sharp dependency to a newer version
The primary reason is that Component Detection
(https://github.com/microsoft/component-detection) pointed out that the
version we used is missing legal information.

Noticing that
https://www.nuget.org/packages/LibGit2Sharp.NativeBinaries/2.0.278/License
shows only a link into the repository, but the full license is shown at
https://www.nuget.org/packages/LibGit2Sharp.NativeBinaries/2.0.322/License
this issue should be hereby resolved.

Besides, it's always good to stay up to date with dependencies.

Since libgit2 is very careful about backwards-compatibility (and
therefore LibGit2Sharp, too), this update should not result in any
change of behavior.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>