scripts/Build.bat: specify nuget pkg install dir
The Build.bat script installs the vswhere package in order to locate
MSBuild if it is not already found on the %PATH%, however it does not
correctly make use of the %VFS_PACKAGESDIR% directory that should
house all the packages used in the build process (such as the
GVFS.ProjFS package).

Correct the script by adding the -OutputDirectory argument to the nuget
install command for vswhere.