Replace x64 check installer with x64compatible
PR #1815 updated the installer dependency version, which seems to have caused the installer to
enforce the AllowedArchitectures=x64 check (which has been there for years). We've received
complaints from users that the latest release refuses to install on their ARM machines, but previous
versions did and worked fine with the default x86/64 emulation.

The latest version of Innosetup adds x64compatible option for the AllowedArchitectures and
ArchitecturesInstallIn64BitMode to address the case of x64 emulation on ARM.