Fix ownership when cloning elevated
On Windows, if the current user is elevated then any directories
created will be owned by the Administrators group. This can cause
problems later on in non-elevated contexts due to git's "dubious
ownership" check. Git for Windows does not currently consider a
non-elevated admin to be the owner of a directory owned by the
Administrators group, though a fix is in progress in the microsoft fork
of git. Libgit2(sharp) also does not have this fix. Also, the GVFS
service which automounts repositories runs under the SYSTEM account,
which would still fail the check even if that fix were in place.

This commit changes the ownership of the .git directory and the working
directory to the current user when cloning.