Update to windows-2025, upgrade to .NET 4.8
window-2019 agent has been deprecated and is no longer available.
The newer agents don't include support for .NET Framework 4.6.1 or the Windows SDK that the native projects targeted.
I tried some workarounds to install support for them as part of the action YML, but failed.
Instead, I've updated the projects to target .NET Framework 4.8 and the latest Windows SDK.