Functional Tests: account for 'partially-hydrated checkout' messages
Let the functional tests account for the "You are in a
partially-hydrated checkout with %d% of tracked files present" message
that was added in Microsoft Git v2.45.2.vfs.0.2.

Do this by filtering out those messages in the same way we do for the
old 'A new version of VFS for Git is available.' upgrader messages, and
the 'Waiting for %s' lock waiting message.

Note that the 'hydration' state message is printed to standard output
and not standard error as the other messages, since this is the output
from the `status` command. We filter both stdout and stderr.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>