TO-DROP: allow attaching conveniently to the functional tests
Currently, it is not possible to run the functional tests from within
Visual Studio, as the environment is set up wrong, somehow. Instead, the
tests have to be run via `script/RunFunctionaltests.bat` (which takes an
optional `--test=<full-class-name>` argument). This makes
single-stepping somewhat hard.

With this commit (which is not intended to be merged into the main
branch!), setting the environment variable `DDD` will force the
functional test to wait until a debugger is attached to the process.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>