ci: avoid failing fast in the matrix jobs
Most of the failures we encounter in VFSforGit's CI runs are flakes. In
those instances, it is not helpful but rather annoying when long-running
jobs are canceled after more than half an hour when they could have run
to completion and succeeded instead. Because when those runs fail, the
first thing I do is to hit "Re-run" to see whether it _is_ a flake.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>