ci: skip also for successful runs for _tree-same_ commits
As far as VFSforGit's CI is concerned, there is no difference between
commits when their _files_ have not changed (it's not like the tip
commit's SHA is persisted in some version number or some such).

Therefore, if there is already a successful run for another commit whose
top-level tree is identical to the one that is about to be tested, skip
the current run in favor of that one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>