ci: when skipping runs, do not skip the required jobs
To increase confidence in Pull Requests, a couple of Checks are
required, for example the PR build must pass.

However, due to one of GitHub Actions' limitations that seem to stay
with us forever, it is impossible to require a specific _workflow_ to
succeed for a PR before it can be merged. You have to specify workflow
_jobs_ that are required.

Therefore, the changes I just made to skip runs when there already
exists a successful run for the same commit (or at least a tree-same
one), would eternally prevent affected PRs from being merged. While the
_workflow_ succeeds in such instances, the (required) workflow _jobs_
would be skipped (and therefore the required checks would not be
fulfilled).

Let's bite the bullet and patch through the information whether to skip
re-running the tests _to the individual jobs_, so that they do run, and
the do get marked as successful, and the Pull Requests can proceed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>