ci: when looking for successful workflow runs, wait for in-progress ones
The Functional Test jobs takes long enough (35 minutes for x86_64, 1
hour 15 minutes for ARM64) that it is worth waiting for them to finish
before kicking off a matrix build that tests the exact same code.

So let's do that, wait for any in-progress run that might result in
some saved time. Granted, this spends cycles of GitHub runners waiting
idly, but GitHub does not provide for a better way.

Note: The `concurrency` attribute _might_ look as helpful in this
context at first sight. But is comically limited: It can only let a
_single_ job wait in a queue, subsequent jobs get canceled (and if you
have notifications turned on, in a loud manner!). And even letting the
job wait would not help here because we also want to handle _tree-same_
commits, not only identical commits, and GitHub does not provide for
a way to identify them in GitHub workflow expressions, either.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>