ci: explicitly restrict `GITHUB_TOKEN` permissions
Yes, yes, CodeQL. It is totally _possible_ that someone with admin
privileges will change the default back to the unsafe `write`
permission. Even if unlikely, let's make it explicit that the `build`
workflow only requires `read` permission. Even in a PR that is only
modifying that workflow, and whose purpose is a totally different one.
Let's conflate separations of concerns. Whatever.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>