Warn-and-continue if gvfs/config download fails
The verbs 'clone', 'mount', 'prefetch', and 'cache-server' each perform
a query to gvfs/config endpoint for two purposes: to get the list of
allowed client versions for the server, and to get the list of
preconfigured cache servers. Currently, if the query fails then the
verb fails.

The most common reason for the query to fail appears to be transient
network/authentication issues, making automount in particular more flaky
than desired.

Mount and Prefetch verbs should always have a cache server configuration
available already, and Clone and Cache-Server both have command-line
options to allow specifying the cache server configuration.

This change modifies the verbs so that if a cache server url is already
available (from local config or command-line option) then the verb will
warn that the version could not be checked and continue instead of
failing if the query to gvfs/config fails.