DEBUG: test runs whose build (but not Functional Tests) succeeded
This is a run whose Functional Test failed:

- https://github.com/microsoft/VFSForGit/actions/runs/16939605701

It failed for legitimate reasons:
https://github.com/microsoft/VFSForGit/actions/runs/16939605701/job/48005131339#step:8:8628
(the Git version that was used forcefully output some warnings).

Anyway, the point of this here commit is not to try to let the
Functional Tests pass, but to test what happens if a previous run was
found whose `build` jobs passed but whose `functional_test` jobs failed:
In that instance, the artifacts must be downloaded from the previous run
and the `build` job must be skipped.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>