Prepare for running the Functional Tests in parallel
This teaches a new command-line option to `GVFS.FunctionalTests.exe`
that allows running only a slice of all tests at a time, via
`--slice=<m>,<n>` where `m` is the zero-based index of the slice
to run, and `n` is the total number of slices to split the tests into.

The idea is to distribute the test cases evenly across the slices,
so that running all slices in parallel will finish in about the same
time as running all tests in a single slice.

This is slightly tricky because the test cases of classes within the
`EnlistmentPerFixture` namespace are known to rely on side effects of
previous test cases within the same class. Therefore, all test cases
of such a class must be run in the same slice.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>