ci: run Functional Tests in parallel
Following the example of git/git, which runs the Windows tests in
parallel to compensate for a very long run time, we split the Functional
Tests into 10 buckets and run them in parallel.

This is particularly useful in light of some flaky tests that frequently
need to be re-run. That way, the cost of re-running a failed test is
reduced by virtue of only having to re-run a _slice_ of the tests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>