ci: do provide the job names expected by the Required PR Checks
In the GitHub UI, it is possible to specify Checks that need to be
successful before a PR can be merged. These "Checks" are actually the
names of jobs in the workflow. It would be much better to specify a
required _workflow_, but only workflow _jobs_ can be specified.

The Required Checks are currently:

  - Functional Tests (Debug, arm64)
  - Functional Tests (Debug, x86_64)
  - Functional Tests (Release, arm64)
  - Functional Tests (Release, x86_64)

So let's make sure that there are job names that match the recorded
names and that indicate a successful workflow run.

An alternative would be to add a single job with a defined name that
depends on all the other jobs, and that can compensate for the missing
GitHub feature to require a whole workflow instead of individual jobs
to have run successfully. But that would interfere with other PRs that
are currently in flight, and will therefore have to wait its turn in
a separate PR.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>