ci: prepare for a single Required Check
GitHub has a feature called "Require status checks before merging":
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/about-protected-branches#require-status-checks-before-merging

What they actually mean by "status checks" is "successful workflow job
runs". And what they _really_ actually mean is "_display labels_ of
successful workflow job runs". That is, workflow job runs are identified
by the label they display.

Example: In VFSforGit, there is currently a "required status check"
that reads "Functional Test (Debug, arm64)". This is the label that is
displayed for the matrix job with vector {configuration: Debug,
architecture: arm64} in the "functional_test" workflow job.

This is quite restrictive! In VFSforGit, specifying those four matrix
job labels is a _work-around_ for the _actual_ requirement, namely that
the workflow defined in `build.yaml` succeeds. It just so happens that
those four matrix jobs are the leaf jobs, i.e. when they all succeed,
the workflow _run_ has succeeded. And vice versa, if the workflow run
failed, at least one of those four matrix jobs must have failed or not
even run.

Now that I multiplied the matrix jobs even further by running the
Functional Tests in parallel, the _display labels_ (and the number) of
the matrix jobs has changed. As a consequence, to appease the "Require
status checks before merging" rule, I had to add _another_ set of
matrix jobs just to guarantee that the same four matrix job labels
exist. This is silly, because those four matrix jobs are not needed at
all for actually testing the code. They are just there to make GitHub
happy.

This commit prepares to change that. It adds a new workflow job
that fits the bill "if the job succeeded, the workflow run must have
succeeded as a hole, and vice versa". This new job will be made the
only "required status check", once this here PR has been merged, and
then we can remove the silly "duplicate" matrix jobs again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>