TryValidateGVFSVersion: allow for the version to include a commit ID
In .NET SDK 8, the Informational Version attribute of the assembly
includes the commit ID of the current HEAD at build time:

https://learn.microsoft.com/en-us/dotnet/core/compatibility/sdk/8.0/source-link

This kind of version might be compliant with SemVer v2.0, but it is
_not_ System.Version-compliant. Therefore, we opted out of this new
behavior _really quickly_ after upgrading from SDK 3 to SDK 9, in order
to fix cloning from repositories where `gvfs/config` requires certain
minimal VFSforGit versions.

In this here commit, we adapt the logic to strip off the commit ID if
present. That way, the version check would continue to work even if we
opted back in to this new behavior that includes the commit SHA in `gvfs
version`'s output.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>