Ensure --no-prefetch clone is honored by maintenance
The first prefetch for a large repository can take some time and (if gvfs.trust-
pack-indexes is false) heavy CPU and disk resources.

If a user cloned with --no-prefetch, let's assume they knew what they were doing
and not automatically prefetch after 15 minutes.

A typical use-case for --no-prefetch is in a CI pipeline which only needs the
head of a branch.