Make ready pack files available during prefetch
When prefetching the commit graph for the first time with
`trust-pack-indexes=false`, the first pack file is the largest and takes
the longest to index. In my test case, it takes ~2 minutes to fetch all
the pack files, ~13 minutes to index the first pack file, and less than
1 minute to index the rest.

A previous pull request parallelized the pack file indexing. This pull
request makes it so as soon as the smaller, more recent pack files are
ready they are moved into the live pack folder and can be accessed by
the user's operations.

A marker file is created when a pack file is moved into the live pack
folder ahead of an older pack file still indexing. If the prefetch is
interrupted, the next prefetch will ignore pack files which have the
marker when determining the timestamp for the prefetch, which ensures
that any missing pack files from the interrupted prefetch are
redownloaded. The marker file is deleted once all the previous pack
files are ready.

After a completed prefetch, any markers (and the pack and index files
they mark) left over from previous incomplete prefetches are deleted.