Fix GetHeadTreeId to only read head commit
The previous implementation of GetHeadTreeId called "git show -s". "-s"
suppresses the diff from the output, but it turns out git still performs
the diff operation internally. The inefficiency isn't necessarily an
issue, but we are running git.exe with the virtualization hook off so
if the objects required for the diff are not already downloaded the
command will fail.

The new implementation produces the same output, but doesn't try to do
a diff at all.